post_mixin_and_factory/pyproject.toml
[tool.poetry]
name = "post-mixin-and-factory"
version = "0.1.0"
description = ""
authors = ["chaeyeon <gkscodus11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-extensions = "^3.2.3"
factory-boy = "^3.3.0"
black = "^24.3.0"
djangorestframework = "^3.15.1"
celery = "^5.3.6"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
